skin_app2/requirements.txt
streamlit==1.39.0
tensorflow==2.16.1
numpy==1.26.4
pandas==2.2.2
Pillow==10.3.0
opencv-python==4.9.0.80
matplotlib==3.8.4
scikit-learn==1.4.2
